start to address review comments, add more error handling

+++ lib/metadata/N2Metadata.hpp
@@ -64,8 +64,8 @@ struct N2MetadataFormat {
     // bin start/end info for convenience
     double bin_start_ERA_deg; /// Earth Rotation Angle at start of bin
     double bin_end_ERA_deg;   /// Earth Rotation Angle at end of bin
-    int64_t bin_start_LAST;   /// local apparent sidereal time (nanoseconds) at start of bin
-    int64_t bin_end_LAST;     /// local apparent sidereal time (nanoseconds) at end of bin
+    double bin_start_LAST;    /// local apparent sidereal time (nanoseconds) at start of bin
+    double bin_end_LAST;      /// local apparent sidereal time (nanoseconds) at end of bin
 
     /// The sequence number of the first FPGA frame integrated into this visibility frame
     uint64_t fpga_start_tick;
